+++ studio/index.ts
@@ -1,12 +1,16 @@
+import { PowerhouseConfig } from '@powerhousedao/config/powerhouse';
 import { Command } from 'commander';
 import fs from 'node:fs';
 import { dirname, isAbsolute, join, resolve } from 'node:path';
-import { startServer } from './server';
+import { startServer, StartServerOptions } from './server';
 
-type PowerhouseConfig = {
-    documentModelsDir?: string;
-    editorsDir?: string;
+export type Project = {
+    name: string;
+    path: string;
 };
+
+const IMPORT_SCRIPT_FILE = 'projects-import.js';
+const projectRoot = process.cwd();
 
 const readJsonFile = (filePath: string): PowerhouseConfig | null => {
     try {
@@ -19,6 +23,24 @@ const readJsonFile = (filePath: string): PowerhouseConfig | null => {
     }
 };
 
+function mapProjects(packages: PowerhouseConfig['packages']): Project[] {
+    if (!packages) {
+        return [];
+    }
+
+    return packages.map(pkg => ({
+        name: pkg.packageName,
+        path: join(
+            projectRoot,
+            'node_modules',
+            pkg.packageName,
+            'dist',
+            'es',
+            'index.js',
+        ),
+    }));
+}
+
 export type ConnectStudioOptions = {
     port?: string;
     host?: boolean;
@@ -27,7 +49,46 @@ export type ConnectStudioOptions = {
     localDocuments?: string;
 };
 
+export function generateImportSctipt(outputPath: string, projects: Project[]) {
+    const importScriptFilePath = join(outputPath, IMPORT_SCRIPT_FILE);
+
+    // create file if it doesn't exist, also create path if it doesn't exist (recursive)
+    if (!fs.existsSync(outputPath)) {
+        fs.mkdirSync(outputPath, { recursive: true });
+    }
+
+    if (!fs.existsSync(importScriptFilePath)) {
+        fs.writeFileSync(importScriptFilePath, '');
+    }
+
+    const imports: string[] = [];
+    const moduleNames: string[] = [];
+    let counter = 0;
+
+    for (const project of projects) {
+        // check if file path exists first
+        if (fs.existsSync(project.path)) {
+            const moduleName = `module${counter}`;
+            moduleNames.push(moduleName);
+            imports.push(`import * as ${moduleName} from '${project.path}';`);
+            counter++;
+        } else {
+            console.error(
+                `Can not import project: ${project.name}, file not found: ${project.path}. Did you build the project or install it?`,
+            );
+        }
+    }
+
+    const exportStatement = `export default [${moduleNames.join(', ')}];`;
+
+    const fileContent = `${imports.join('\n')}\n\n${exportStatement}`;
+
+    fs.writeFileSync(importScriptFilePath, fileContent);
+}
+
 export function startConnectStudio(options: ConnectStudioOptions) {
+    let serverOptions: StartServerOptions = {};
+
     if (options.port) {
         process.env.PORT = options.port;
     }
@@ -42,6 +103,20 @@ export function startConnectStudio(options: ConnectStudioOptions) {
 
         const configFileDir = dirname(options.configFile);
 
+        if (config.packages && config.packages.length > 0) {
+            const packages = mapProjects(config.packages);
+            generateImportSctipt(configFileDir, packages);
+            process.env.LOAD_EXTERNAL_PROJECTS = 'true';
+
+            serverOptions = {
+                enableExternalProjects: true,
+                projectsImportPath: resolve(configFileDir, IMPORT_SCRIPT_FILE),
+            };
+        } else {
+            process.env.LOAD_EXTERNAL_PROJECTS = 'false';
+            serverOptions = { enableExternalProjects: false };
+        }
+
         if (config.documentModelsDir) {
             process.env.LOCAL_DOCUMENT_MODELS = isAbsolute(
                 config.documentModelsDir,
@@ -55,6 +130,21 @@ export function startConnectStudio(options: ConnectStudioOptions) {
                 ? config.editorsDir
                 : join(configFileDir, config.editorsDir);
         }
+
+        if (config.studio?.port) {
+            process.env.PORT = config.studio.port.toString();
+        }
+
+        if (typeof config.studio?.openBrowser === 'boolean') {
+            process.env.OPEN_BROWSER = config.studio.openBrowser.toString();
+        }
+
+        if (config.studio?.host) {
+            process.env.HOST = config.studio.host;
+        }
+    } else {
+        process.env.LOAD_EXTERNAL_PROJECTS = 'false';
+        serverOptions = { enableExternalProjects: false };
     }
 
     if (options.localEditors) {
@@ -65,7 +155,7 @@ export function startConnectStudio(options: ConnectStudioOptions) {
         process.env.LOCAL_DOCUMENT_MODELS = options.localDocuments;
     }
 
-    return startServer().catch(error => {
+    return startServer(serverOptions).catch(error => {
         throw error;
     });
 }